Updated nav_bar.tsx
Used a class to further modify the Image's width, and made its width prop exactly the full size it can go (when paired with w-full).

Changed links flexbox from space-around to space-between. Added vertical padding so that it is easier to click the links. Changed the size of the link text when the screen size is less than XL.