Updated page.tsx
Changed the image to be always absolute. It will have less opacity for any screen size that is smaller than the LG size. Its height will now be auto and its width full if the screen becomes smaller than the MD size.

Changed the first div to only be displayed as a grid if the screen is of LG size or greater. Removed the grid-cols-2 styling.