Updated nav_bar.tsx, page.tsx
In page.tsx: Placed the first div back into a one-column grid so that the items can align to the bottom of the grid and the height can be 100vh.

In nav_bar.tsx: Being absolute is now only for screens with a size of LG or larger. If the screen is not such a size, then the navigation bar becomes part of the normal flow and has a margin instead of an offset.